+++ src/main/java/nl/_42/restsecure/autoconfigure/authentication/mfa/MfaSetupService.java
@@ -5,4 +5,8 @@ public interface MfaSetupService {
     String generateSecret();
 
     String generateQrCode(String secret, String label) throws MfaException;
+    
+    default void setupEmailMfa(String email) {
+        // Default implementation does nothing, can be overridden by implementers
+    }
 }