+++ src/main/java/nl/_42/restsecure/autoconfigure/authentication/mfa/MfaType.java
@@ -1,0 +1,7 @@
+package nl._42.restsecure.autoconfigure.authentication.mfa;
+
+public enum MfaType {
+    NONE,
+    TOTP,
+    EMAIL
+}