+++ src/main/java/nl/_42/restsecure/autoconfigure/authentication/mfa/email/EmailCodeService.java
@@ -1,0 +1,6 @@
+package nl._42.restsecure.autoconfigure.authentication.mfa.email;
+
+public interface EmailCodeService {
+    void generateAndSendCode(String email);
+    boolean verifyCode(String email, String code);
+}